requirements.txt
streamlit
openai
pdfplumber
python-dotenv
fpdf
PyPDF2
google-generativeai
cohere